gen.sh, kigen.py, replace.sh made so it is callable in any directory version lookup table added templates moved under version folders projects moved under version folders

+++ scripts/gen.sh
@@ -1,4 +1,14 @@
+thisPath=$(realpath "$0")
 req=$1
+thisDirPath=$(dirname "$thisPath")
+prjDirPath=$thisDirPath/..
+
+cd $prjDirPath
+if [[ $? != 0 ]]
+then
+    echo cd to project directory failed
+    exit 1
+fi
 
 # clear old files
 rm -rf ./out
